+++ CHANGELOG.md
@@ -5,9 +5,12 @@
 ### Changed
 
 - Update Bitmovin's native Android SDK version to `3.145.0+jason`
+- Android/iOS: Keep `PlayerView`'s `config` property immutable, except for `pictureInPictureConfig` which can be updated dynamically after initialization
 
 ## [1.11.0] - 2026-03-05
+
 ### Added
+
 - Support for streaming ads with the Bitmovin Advertising Module (BAM) via `AdSourceType.BITMOVIN` in `AdvertisingConfig` ad schedules. Note: on iOS, VMAP is currently not supported with BAM.
 
 ### Changed